+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Bug Fixes
 - Search works properly when library is updated or albums are expanded
 - Fixed adding/removing buffers in vertical orientation
 - Loved ordering no longer reintroduces filtered-out albums
+- Cycling through buffers works when the next buffer won't accept focus
 
 
 Version 0.2.1